restructure package
add sub-package structures and simulations
rename modules and regroup classes into modules
