+++ udkm1Dsim/__init__.py
@@ -2,18 +2,14 @@ from pint import UnitRegistry
 u = UnitRegistry()
 u.default_format = '~P'
 Q_ = u.Quantity
-from .atoms import Atom, AtomMixed
-from .layer import Layer, AmorphousLayer, UnitCell
-from .structure import Structure
-from .simulation import Simulation
-from .heat import Heat
-from .phonon import Phonon
-from .phononNum import PhononNum
-from .magnetization import Magnetization
-from .xray import Xray
-from .xrayKin import XrayKin
-from .xrayDyn import XrayDyn
-from .xrayDynMag import XrayDynMag
+from .structures.atoms import Atom, AtomMixed
+from .structures.layers import Layer, AmorphousLayer, UnitCell
+from .structures.structure import Structure
+from .simulations.simulation import Simulation
+from .simulations.heat import Heat
+from .simulations.phonons import Phonon, PhononNum
+from .simulations.magnetization import Magnetization
+from .simulations.xrays import Xray, XrayKin, XrayDyn, XrayDynMag
 
 __all__ = ['Atom', 'AtomMixed', 'Layer', 'AmorphousLayer', 'UnitCell', 'Structure',
            'Simulation', 'Heat', 'Phonon', 'PhononNum', 'Magnetization', 'Xray',
