+++ udkm1Dsim/structures/atoms.py
@@ -22,15 +22,15 @@
 # OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
 # OR OTHER DEALINGS IN THE SOFTWARE.
 
-__all__ = ["Atom", "AtomMixed"]
+__all__ = ['Atom', 'AtomMixed']
 
-__docformat__ = "restructuredtext"
+__docformat__ = 'restructuredtext'
 
+from .. import u, Q_
 import os
 import numpy as np
 import scipy.constants as constants
 from tabulate import tabulate
-from . import u, Q_
 
 
 class Atom:
@@ -95,7 +95,7 @@ class Atom:
 
         try:
             filename = os.path.join(os.path.dirname(__file__),
-                                    'parameters/elements.dat')
+                                    '../parameters/elements.dat')
             symbols = np.genfromtxt(filename, dtype='U2', usecols=(0))
             elements = np.genfromtxt(filename, dtype='U15, i8, f8', usecols=(1, 2, 3))
             [rowidx] = np.where(symbols == self.symbol)
@@ -146,7 +146,7 @@ class Atom:
         """
         if not filename:
             filename = os.path.join(os.path.dirname(__file__),
-                                    'parameters/atomic_form_factors/chantler/{:s}.cf'.format(
+                                    '../parameters/atomic_form_factors/chantler/{:s}.cf'.format(
                                             self.symbol.lower()))
         try:
             f = np.genfromtxt(filename, skip_header=0)
@@ -195,7 +195,7 @@ class Atom:
 
         """
         filename = os.path.join(os.path.dirname(__file__),
-                                'parameters/atomic_form_factors/cromermann.txt')
+                                '../parameters/atomic_form_factors/cromermann.txt')
         try:
             cm = np.genfromtxt(filename, skip_header=1,
                                usecols=(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11))
@@ -281,7 +281,7 @@ class Atom:
         """
         if not filename:
             filename = os.path.join(os.path.dirname(__file__),
-                                    'parameters/magnetic_form_factors/{:s}.mf'.format(
+                                    '../parameters/magnetic_form_factors/{:s}.mf'.format(
                                             self.symbol))
         try:
             f = np.genfromtxt(filename)
